Bot infrastructure to a separate project.
