+++ BotInfrastructure/Interface/Command/RegexCommand.cs
@@ -1,0 +1,8 @@
+﻿using System.Text.RegularExpressions;
+
+namespace BotInfrastructure.Interface.Command;
+
+public abstract class RegexCommand(string name, string validRegexArg) : MultipleRegexCommand(name, [validRegexArg])
+{
+	public Regex ValidRegex => ValidRegexes[0];
+}
