+++ BotInfrastructure/Interface/IBot.cs
@@ -1,0 +1,6 @@
+﻿namespace BotInfrastructure.Interface;
+
+public interface IBot
+{
+	Task StartAsync(CancellationToken cancelToken);
+}
