+++ BotInfrastructure/Model/CommandConfiguration.cs
@@ -1,0 +1,7 @@
+﻿namespace BotInfrastructure.Model;
+
+public class CommandConfiguration
+{
+	public required Dictionary<string, string> Names { get; set; } = new();
+	public required string DateFormat { get; set; }
+}
