+++ BotInfrastructure/commands.txt
@@ -1,0 +1,3 @@
+﻿dynam - График курса за большие периоды
+rate - Курс на любую дату или любой день текущего месяца
+conv - Конвертировать число в валюту на любую дату или день
