+++ FxBot.sln
@@ -1,11 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31112.23
+# Visual Studio Version 17
+VisualStudioVersion = 17.10.35013.160
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FxBot", "FxBot\FxBot.csproj", "{1CEFBBCF-162A-4949-AF9B-BF437252E50C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QuoteService", "QuoteService\QuoteService.csproj", "{713D39DE-7525-4CD0-BFE9-71456CA6619F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuoteService", "QuoteService\QuoteService.csproj", "{713D39DE-7525-4CD0-BFE9-71456CA6619F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BotInfrastructure", "BotInfrastructure\BotInfrastructure.csproj", "{A18DDAA1-5927-43DF-82A8-48065C427613}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{713D39DE-7525-4CD0-BFE9-71456CA6619F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{713D39DE-7525-4CD0-BFE9-71456CA6619F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{713D39DE-7525-4CD0-BFE9-71456CA6619F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A18DDAA1-5927-43DF-82A8-48065C427613}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A18DDAA1-5927-43DF-82A8-48065C427613}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A18DDAA1-5927-43DF-82A8-48065C427613}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A18DDAA1-5927-43DF-82A8-48065C427613}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
