+++ FxBot/ConfigurationExtensions.cs
